% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansbach-round-scheme.R
\name{ansbach_round_scheme}
\alias{ansbach_round_scheme}
\title{HS Ansbach rounding scheme for grades}
\usage{
ansbach_round_scheme()
}
\value{
tibble
}
\description{
At HS Ansbach the grading scheme is like 1, 1.3, 1.7, 2.0, etc. (but no "1.5" etc)
}
\examples{
`ansbach_round_scheme()`


}
