% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_grade.R
\name{assign_grade}
\alias{assign_grade}
\title{Assign Grades}
\usage{
assign_grade(d, grading_scheme = NULL, var, failing_grades = c("5"))
}
\arguments{
\item{d}{data frame}

\item{grading_scheme}{which grading scheme to use}

\item{var}{variable for which a grade should be assigned}

\item{failing_grades}{grades when student is failing}

\item{number_of_grade}{number of grades that should be in the grading_scheme}
}
\value{
data frame with grades (column) added
}
\description{
Given a variable with points, percentage or other, a grade is
assigned given a grading scheme which acts as a function ascribing the
mapping from proints to grade
}
\details{
The dataframe \code{d} is expected to have a variable \code{var} which provides the
<<<<<<< HEAD
performance data to be graded.
=======
performance data to be graded (not only one student, but a column of students).
>>>>>>> e10612fcb8bac6580dc6942a9582d60622e57b27
\code{grading_scheme} is a named numeric vector such as "\code{4} == .5", etc.
\code{failing_grades} is defaulted with \code{c("5", "4.3", "4.7")}.
\code{number_of_grade} is defaulted with \code{11} (ie., grades from 1.0, 1.3, ...4, 5)
\code{var} is defaulted with \code{"correct_prop"}
}
\examples{
assign_grade(mtcars, var = "am")

}
