% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonuspoints.R
\name{bonuspoints}
\alias{bonuspoints}
\title{bonuspoints}
\usage{
bonuspoints(item_vec, return_type = "logical")
}
\arguments{
\item{item_vec}{vector of 0/1 values (numeric)}

\item{return_type}{type of return vector}
}
\value{
vector indicating wich students get a bonuspoint
}
\description{
Gives 1 bonuspoint to exercises which have been \emph{wrongly} answered
}
\examples{
item_vec <- c("a" = 1, "b" = 0, "c" = 0, "d" = 1)
bonuspoints(item_vec)
bonuspoints(item_vec, return_type = "01")
bonuspoints(item_vec, return_type = "02")
}
