% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_adj_set_answers.R
\name{build_adj_set_answers}
\alias{build_adj_set_answers}
\title{build adjustment set answers}
\usage{
build_adj_set_answers(dag_def, exposure_var, outcome_var)
}
\arguments{
\item{dag_def}{dag defintion, based on R package dagitty (string)}

\item{exposure_var}{which variable is the exposure (string)?}

\item{outcome_var}{which var. is the outcome variable? (string)}
}
\value{
tibble wit 1 correct and 4 incorrect solutions
}
\description{
Producing single choice exercises for minimal adjustment sets
}
\details{
This function produces a single choice answer list for suitable
for R/exam exercises. Use this function to build single choice question asking for the minimal adjustment set
}
\examples{
\dontrun{build_adj_set_answers(my_dag, x, y,)}
}
