% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_answer_options.R
\name{build_answer_options}
\alias{build_answer_options}
\title{Build answer options}
\usage{
build_answer_options(
  correct_answer,
  nr_answer_options = 5,
  algo_wrong_answers = "prop",
  scale_factor = 2,
  format_answer = FALSE
)
}
\arguments{
\item{correct_answer}{numeric the value of the correct answer}

\item{nr_answer_options}{defaults to 5 (integer), other values or not supported currently}

\item{algo_wrong_answers}{"coef" (large range of values) or "prop" (values from 0 to 1)}

\item{scale_factor}{defaults to 2, useful for building plausible distractors which as a function of a multiple of the true solution}

\item{format_answer}{defaults to FALSE using the functions exams::fmt()}
}
\value{
numeric vector of length 5
}
\description{
Buils answer options for multiple choice questions
}
\details{
This functions builds 5 (and only 5, currently) dynamic answer options.
Useful for dynamic exercises in r-exams.
At present, two algorithms are supported for creating answer options: "coef" and "prop"
"coef" builds options by multiplying the correct answer by \code{scale_factor} (and adds a constant in some instances; see below for details)
"prop" build options by multiplying the odds by some \code{scale_factor} (3 in default)
The correct answer is the FIRST of element of the returned answers (the rest are the wrong ones)
}
\examples{
\dontrun{build_answer_options(42)}
 42.0000000 -41.0000000   0.2896458   0.4603042   0.7941220
}
