% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_studentilist.R
\name{build_studentilist}
\alias{build_studentilist}
\title{Build tidy Studenti list}
\usage{
build_studentilist(inputfile, outputfile)
}
\arguments{
\item{inputfile}{input xlsx file}

\item{outputfile}{output csv file}
}
\value{
tidy data frame with studenti names etc.
}
\description{
This function takes a primuss export consisting of names, matrikelnummers and email adresses
of students. Then a more well-behaved csv file is constructed.
}
\examples{
inputfile <- "/Users/sebastiansaueruser/Google Drive/ORGA-HS-Ansbach/22-WS/AWM Studierende_Email & MATR_ Stand 24.01.23.xlsx"
outputfile <- "/Users/sebastiansaueruser/Google Drive/ORGA-HS-Ansbach/22-WS/studentis-awm-2022.csv"
build_studentilist(inputfile, outputfile)
}
