% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonicalize_df.R
\name{canonicalize_df}
\alias{canonicalize_df}
\title{Canonicalize a data frame}
\usage{
canonicalize_df(df, output_var, keep_only_numeric = TRUE, verbose = TRUE)
}
\arguments{
\item{df}{Data frame to be canonicalized (df)}

\item{output_var}{the dependent variable of a (possibly) intended model (string)}

\item{keep_only_numeric}{should only numeric variables be kept? (lgl), defaults to TRUE}
}
\value{
data frame canonicalized
}
\description{
Converts a data frame to a standard form with 1 output variable
and two predictors, all with standard names (av, uv1, uv2).
Currently, only 2 predictors are supported.
}
\examples{
\dontrun{canonicalize_df(my_df, my_outputvar)}
}
