% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam2yamlrmd.R
\name{exam2yamlrmd}
\alias{exam2yamlrmd}
\title{exam2yamlrmd}
\usage{
exam2yamlrmd(
  examfile,
  path_output = "",
  ex_sol_str = c("Aufgabe", "Lösung", "Categories"),
  print_categories = FALSE,
  separate_ex_sol = rep("</br>", 10),
  more_yaml = NULL,
  verbose = TRUE,
  header_level = 1,
  file_suffix = ".Rmd"
)
}
\arguments{
\item{examfile}{path to input file in r-exam format (string)}

\item{path_output}{output path (string)}

\item{ex_sol_str}{wording for the parts of the exercise (useful for translation purposes)}

\item{print_categories}{should the yaml field 'categories' be printed?}

\item{separate_ex_sol}{How should the question and solution be visually separated?}

\item{more_yaml}{more yaml to add which was not given in the r-exam file}

\item{verbose}{talkative, yes or no? (TRUE or FALSE)}

\item{header_level}{how many '#' should be padded in order to indicate the header leve?}

\item{file_suffix}{defaults to ".Rmd" (string), other useful option: ".qmd"}
}
\value{
writes an markdown file
}
\description{
Converts r-exam File to yaml-headed RMD file
}
\details{
This function calls parse_examfile() using \code{examfile} as input and based on the output it calls then write_yamlrmdfile().
}
\examples{
\dontrun{exam2yamlrmd(examfile, path_output)}
examfile_name <- "/Users/sebastiansaueruser/github-repos/rexams-exercises/exercises/sebastiansauer/de/Bayes/pigs2.Rmd"
out <- exam2yamlrmd(
 examfile = examfile_name,
 path_output = "/Users/sebastiansaueruser/Downloads")





}
