% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_grades.R
\name{finalize_grades}
\alias{finalize_grades}
\title{finalize grades}
\usage{
finalize_grades(
  d_error,
  thresholds,
  grades_particulars = NULL,
  save_to_disk = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{d_error}{df with error per student. Output of \code{comp_error_submissions}}

\item{thresholds}{thresholds for the grades}

\item{grades_particulars}{individual changes to the grades (df)}

\item{save_to_disk}{save as \code{d_grades.csv} and \code{d_grades.rds}, if TRUE (lgl)}

\item{verbose}{more output? (lgl)}

\item{...}{}
}
\value{
df with grdades (notenliste)
}
\description{
Given the predictive error per student, compute the individual grades.
}
\examples{
\dontrun{finalize_grades(my_d_error, my_grades_templates, my_thresholds,)}
}
