% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Get Datasets from Vincent Arel Bundock's page}
\usage{
get_datasets(d_path = NULL, d_filter = NULL)
}
\arguments{
\item{d_path}{CSV file with data set information}

\item{d_filter}{Name of Package for filter for}
}
\value{
data frame with data set information
}
\description{
Downloads all data frames from a source page (default: incent Arel Bundock's page),
according to a filter (i.e., a package to filter for)
}
\examples{
out <- get_datasets(d_filter = "openintro")
}
