% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-funs.R
\name{get_grade_scheme}
\alias{get_grade_scheme}
\title{get_grade_scheme}
\usage{
get_grade_scheme(scheme = 1)
}
\arguments{
\item{scheme}{which scheme (int)? Currently only 1 scheme is implemented}
}
\value{
scheme (numeric)
}
\description{
Get grade scheme as numeric vector
}
\details{
Returns one (of possibly several) grading scheme
}
\examples{
\dontrun{get_grade_scheme()}
}
