% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pckgs.R
\name{get_pckgs}
\alias{get_pckgs}
\title{Get List of R Packages Ised in a Project}
\usage{
get_pckgs(pattern = "\\\\.qmd$|\\\\.R$|\\\\.Rmd$", recursive = TRUE)
}
\arguments{
\item{pattern}{(chr)}

\item{recursive}{(lgl)}
}
\value{
character vector with identified R packages in the project
}
\description{
Parses all files ending with "qmd", "R" and "Rmd".
Returns a string of length n with all n packages identified.
Note that only calls with "library()" and "require()" will be detected.
}
\examples{
get_pckgs()
}
