% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_em_all.R
\name{grade_em_all}
\alias{grade_em_all}
\title{Grade em all}
\usage{
grade_em_all(
  data_path_moodle_csv,
  grading_scheme = NULL,
  verbose = TRUE,
  n_bonus_points = 0,
  suppress_details = TRUE,
  suppress_warnings = TRUE,
  comments = NULL,
  ...
)
}
\arguments{
\item{data_path_moodle_csv}{A csv file, possible coming from moodle with exam items. See details.}

\item{grading_scheme}{Defines thresholds for grades, defaults to  \code{c(51, 55, 60, 65, 70, 75, 80, 85, 90, 95)}}

\item{verbose}{if true, additional processing infos are printed (lgl, defaults to TRUE)}

\item{n_bonus_points}{Be kind to your students (int, defaults to zero)}

\item{suppress_details}{suppresses the output of additional columns (lgl, defaults to TRUE)}

\item{comments}{comments passed on to each student (chr, defaults to NULL)}

\item{...}{additional arguments, currently not used}
}
\value{
csv file with grades
}
\description{
Assigns grades to exams
}
\details{
This function assigns a grade to each student (row)
drawing upon the exam items in the input csv moodle file.
Some preprocssing is being done, such as cleaning the col names.
}
\examples{
\dontrun{grade_em_all(data.csv)}

data_path_moodle_csv <- "/Users/sebastiansaueruser/Google Drive/Lehre/Lehre_AKTUELL/2022-SoSe/QM2/Prüfung/Ergebnisse/raw-data/AWM-QM2-SoSe22-Prüfungsaufgaben-Bewertungen (2).csv"
my_grading_scheme <-  c(45, 50, 55, 60, 65, 70, 75, 80, 85, 90)
out <- grade_em_all(data_path_moodle_csv, my_grading_scheme)

}
