% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_essay.R
\name{grade_essay}
\alias{grade_essay}
\title{grade essay}
\usage{
grade_essay(d)
}
\arguments{
\item{d}{data frame with raw grading data}
}
\value{
vector (num) with graded data
}
\description{
takes a dataframe with 4 columns and grades it
}
\details{
Expects the followings columns: Methodik, Formales, Inhalt, Bonus
(num/num/num/chr).
}
