% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade-percentage-points.R
\name{grade_percentage_points}
\alias{grade_percentage_points}
\title{Grade percentage points}
\usage{
grade_percentage_points(
  data_path_moodle_csv,
  verbose = TRUE,
  n_bonus_points = 0
)
}
\arguments{
\item{data_path_moodle_csv}{path to csv file with moodle test results (stringr)}

\item{verbose}{print additional information (lgf, defaults to TRUE)}

\item{n_bonus_points}{how many bonus points should be granted (integer, defaults to zero)}
}
\value{
tibble with percentage points grades
}
\description{
Assigns percentage points to test results
}
\examples{
\dontrun{grade_percentage_points(my_csv_file)}
}
