% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_scheme.R
\name{grading_scheme}
\alias{grading_scheme}
\title{Grading scheme}
\usage{
grading_scheme(scheme_type = 1, output_type = "proportion")
}
\arguments{
\item{scheme_type}{which grading scheme (integer, defaults to 1)}

\item{output_type}{scheme in percentage points or as proportion, choose either "percentage" or "proportion")}
}
\value{
numeric named vector of varying length depending on type
}
\description{
Returns various grading schemes
}
\examples{
grading_scheme()
5   4 3.7 3.3 3.0 2.7 2.3 2.0 1.7 1.3 1.0
0  50  55  60  65  70  75  80  85  90  95
}
