% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading_scheme2.R
\name{grading_scheme2}
\alias{grading_scheme2}
\title{Grading scheme 2}
\usage{
grading_scheme2(min_pass = 50, output_type = "proportion")
}
\arguments{
\item{min_pass}{minimal percentage points to pass the exam}

\item{output_type}{scheme in percentage points or as proportion, choose either "percentage" or "proportion")}
}
\value{
named numeric vector of length 12
}
\description{
Returns various grading schemes
}
\examples{
grading_scheme2()
5    4  3.7  3.3  3.0  2.7  2.3  2.0  1.7  1.3  1.0
0.00 0.50 0.55 0.60 0.65 0.70 0.75 0.80 0.85 0.90 0.95
}
