% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_gradinglists.R
\name{join_gradinglists}
\alias{join_gradinglists}
\title{Join two grading lists}
\usage{
join_gradinglists(moodle_grading_d, course_file, comments = NULL)
}
\arguments{
\item{moodle_grading_d}{data frame with results of moodle grading}

\item{course_file}{path to XLSX file from institatution}

\item{comments}{comments passed to each student (separate comments are not supported)}
}
\value{
data frame for grading based on the instutation name list
}
\description{
Join moodle grading list with Instituation's grading list
}
\details{
Based on the Moodle csv file, a grading list has been created.
This function now merges the moodle grading list with a XLSX template file
from the institatution. Merging (joining) is performed by based on surname
and first names of the student. Joining errors are reported.
}
\examples{
 \dontrun{join_gradinglists(d1, d2)}
}
