% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_gradinglists2.R
\name{join_gradinglists2}
\alias{join_gradinglists2}
\title{Join two grading lists}
\usage{
join_gradinglists2(moodle_grading_d, course_d, grades_var = "grade")
}
\arguments{
\item{moodle_grading_d}{data frame with results of moodle grading}

\item{course_d}{data frame without grades from university}

\item{grades_var}{variable in Moodle table holding the grades of the students}

\item{comments}{comments passed to each student (separate comments are not supported)}
}
\value{
data frame for grading based on the instutation name list
}
\description{
Join moodle grading list with Instituation's grading list
}
\details{
Based on the Moodle csv file, a grading list has been created.
This function now merges the moodle grading list with a XLSX template file
from the institution. Merging (joining) is performed by based on surname
and first names of the student. Joining errors are reported. Note that
names are no unique nor save way to join lists.
}
\examples{
 \dontrun{join_gradinglists(d1, d2)}
}
