% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_name.R
\name{lookup_name}
\alias{lookup_name}
\title{lookup_name}
\usage{
lookup_name(csv_file, student_id)
}
\arguments{
\item{csv_file}{file with ids and names of students}

\item{student_id}{unique ID (Matrikelnummer)}
}
\value{
name of student
}
\description{
Looks up the name of a student given an ID (Matrikelnummer)
}
\details{
Given a csv file and a student id (Matrikelnummer),
the name of the person (student) is returend
}
\examples{
\dontrun{lookup_name(my_students, "12345678")}
}
