% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_split_all.R
\name{median_split_all}
\alias{median_split_all}
\title{Median split for all variables}
\usage{
median_split_all(data)
}
\arguments{
\item{data}{}
}
\value{
dataframe with median splitted variables as factors
}
\description{
Splits all variables at the median.
Expects numeric variables only.
Split variables are factors with the levels "high" and "low".
}
\examples{
median_split_all(mtcars)
}
