% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notenstufen.R
\name{notenstufen}
\alias{notenstufen}
\title{Notenstufen}
\usage{
notenstufen(type = 1)
}
\arguments{
\item{type}{type of notenstufen, currently are 4 types supported (integer, defaults to 1)}
}
\value{
numeric vector, can be of different length depending on type
}
\description{
Returns notenstufen of various types (e.g, "1.0", "1.3",...)
}
\details{
Teh different types support either coarse grades (1,2,3,...) or more fine grained grades (1.3, 1.7, ...)
}
\examples{
notenstufen()

}
