% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_examfile.R
\name{parse_examfile}
\alias{parse_examfile}
\title{parse_examfile}
\usage{
parse_examfile(examfile, verbose = FALSE)
}
\arguments{
\item{examfile}{path to r-exam exercise file. Can be remote or local.}

\item{verbose}{Print additional information?}
}
\value{
list with the parts of the exercise file
}
\description{
Parse r-exams exercise file
}
\details{
Parses the follwing parts of an r-exam file: 1. metadata, 2. pre-question part,
3. question, 4. solution. The metadata will be converted to regular yaml.
}
\examples{
\dontrun{parse_examfile(myexamfile)}
library(teachertools)
examfile_name <- "/Users/sebastiansaueruser/github-repos/rexams-exercises/exercises/sebastiansauer/de/Bayes/pigs2.Rmd"

fileparsed <- parse_examfile(examfile_name)
fileparsed
}
