% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_FUNS.R
\name{parse_first_names}
\alias{parse_first_names}
\title{parse_first_names}
\usage{
parse_first_names(submission)
}
\arguments{
\item{submission}{string with submission data (name, student id)}
}
\value{
first name of student
}
\description{
Takes submission meta data and returns student's first name
}
\details{
Give a string with the student meta data, the first name of the student is returned
}
\examples{
\dontrun{parse_first_names(submission1)}
}
