% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_FUNS.R
\name{parse_last_names}
\alias{parse_last_names}
\title{parse_last_names}
\usage{
parse_last_names(submission)
}
\arguments{
\item{submission}{string with submission data (name, student id)}
}
\value{
last name of student
}
\description{
Takes submission meta data and returns student's last name
}
\details{
Give a string with the student meta data, the last name of the student is returned
}
\examples{
\dontrun{parse_last_name(submission1)}
}
