% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_FUNS.R
\name{parse_matrikelnummer}
\alias{parse_matrikelnummer}
\title{parse_matrikelnummer}
\usage{
parse_matrikelnummer(submission, n_digits = 8)
}
\arguments{
\item{submission}{string with submission data (name, student id)}
}
\value{
matrikelnummer of student
}
\description{
Takes submission meta data and returns student's matrikelnummer (ID number)
}
\details{
Given a string with the student meta data, the matrikelnummer of the student is returned
It is assumed (per default) that the ID is a number of 8 digits
}
\examples{
\dontrun{parse_matrikelnummer(submission1)}
}
