% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_names_raw.R
\name{parse_names_raw}
\alias{parse_names_raw}
\title{parse students names returns names and matrikelnummers of raw submissions files}
\usage{
parse_names_raw(paths)
}
\arguments{
\item{name}{description path path to submissions raw file from moodle}
}
\value{
sanitized data frame of individual submission. See details.
}
\description{
parse students names returns names and matrikelnummers of raw submissions files
}
