% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grade_distribution.R
\name{plot_grade_distribution}
\alias{plot_grade_distribution}
\title{plot_grade_distribution}
\usage{
plot_grade_distribution(d_grades)
}
\arguments{
\item{d_grades}{df with grades, particularly a column \code{grade} containing the grades}
}
\value{
ggplot2 object
}
\description{
Plots Distribution of Grades
}
\examples{
\dontrun{plot_grade_distribution(d_grades)}
}
