% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nominal_association.R
\name{plot_nominal_association}
\alias{plot_nominal_association}
\title{Visualize a nominal association of a two-by-two dataset}
\usage{
plot_nominal_association(data, var1, var2, label1 = NULL, label2 = NULL)
}
\arguments{
\item{data}{dataframe with numeric variables}

\item{var1}{character, name of variable going to x axis}

\item{var2}{character, name of variable to fill}

\item{label1}{label for var1 (character)}

\item{label2}{label for var2 (character)}
}
\value{
ggplot object (bar graph of two-by-two assocation)
}
\description{
Expects a dataset with two numeric variables,
which are median splitted (using \code{median_split_all}).
Next, a bar graph is plotted and returned.
}
\examples{
plot_nominal_association(mtcars, "mpg", "hp")
plot_nominal_assocation(d, "bill_length_mm", "bill_depth_mm")
}
