% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_notenschwellen.R
\name{plot_notenschwellen}
\alias{plot_notenschwellen}
\title{Plot Notenschwellen}
\usage{
plot_notenschwellen(scheme = grading_scheme2())
}
\arguments{
\item{scheme}{grading scheme such as given by \code{grading_scheme2()}}
}
\value{
ggplot2 object
}
\description{
Assigns minimal proportion of points needed for a given grade, plot this table then
}
\examples{
plot_notenschwellen()
}
