% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_notenspiegel.R
\name{plot_notenspiegel}
\alias{plot_notenspiegel}
\title{Plot Notenspiegel}
\usage{
plot_notenspiegel(data, plot_title = "Notenverteilung")
}
\arguments{
\item{data}{containing grades}

\item{plot_title}{Title to be assigned to the plot}
}
\value{
ggplot2 object
}
\description{
Plot Grade Distribution (Notenspiegel)
}
\details{
Plot distribution of grades, a simple wrapper around ggplot2
Annotations are currently in German language
}
\examples{
\dontrun{plot_notenspiegel(grades_data)}
}
