% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_primes.R
\docType{data}
\name{primes}
\alias{primes}
\title{The first 1000 Prime numbers}
\format{
data frame with 2 columns and 1000 rows

\describe{
\item{id}{integer from 1 to 1000}
\item{prime_number}{prime number (integer)}
}
}
\usage{
data(primes)
}
\description{
Table with the first 1000 prime numbers
}
\keyword{datasets}
