% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_grade.R
\name{quick_grade}
\alias{quick_grade}
\title{quickly grade a prediction contest submission}
\usage{
quick_grade(df_estimate, df_truth, name_outcome_variable)
}
\arguments{
\item{df_estimate_path}{path and file name of estimate csv file (character)}

\item{df_truth_path}{path and file name of truth csv file (character)}

\item{name_outcome_var}{name of outcome variable (character)}
}
\value{
numeric value of rmse performance (tibble)
}
\description{
Note that no sanitizing of the csv file is taking place.
All that happens is that truth and estimate csv files are loaded
and then their correspondence as to RMSE is returned.
Note that "id" and "pred" are the assumed columns in the estimate csv.
An "id" column ranging from 1 to nrow() is added to the truth csv.
The name of the outcome variable column must be supplied.
}
\examples{
quick_grade(estimate_file, truth_file, outcome_var)
}
