% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_exs.R
\name{render_exs}
\alias{render_exs}
\title{render R/exams exercises}
\usage{
render_exs(
  exs,
  thema_nr = "",
  render_moodle = FALSE,
  render_html = TRUE,
  render_pdf_print = TRUE,
  render_markdown = TRUE,
  render_yamlrmd = FALSE,
  output_path,
  template = "plain.html",
  use_seed = TRUE,
  n_students = 1,
  output_name = NULL,
  prime_factor = 1,
  my_edir = "exercises/sebastiansauer",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{exs}{list of file name of exercises to be rendered (string)}

\item{thema_nr}{(optinal) Should a number be suffixed to the resulting exercise file? (string)}

\item{render_moodle}{(FALSE) render to Moodle format?}

\item{output_path}{path where should the output file be saved (string)}

\item{template}{the HTML template to use (path) (string)}

\item{use_seed}{(TRUE) Should the random seed for the exercises (if used) be stored for reproducibility?}

\item{n_students}{(1) How many variants should be divised in case of dynamic exercises?}

\item{output_name}{(NULL, optional) an optional output name of the resulting file}

\item{prime_factor}{(1) use different primes for the seeds}

\item{my_edir}{path where the exercise file can be found (string)}

\item{verbose}{(FALSE) print out details}

\item{html}{(TRUE) render to HTML format using a simple stylesheet?}

\item{pdf_print}{(TRUE) print from HTML to pdf?}

\item{yamlrmd}{(FALSE) render to yaml rmd?}
}
\value{
nothing of importance, but files are written
}
\description{
Renders R/exams exercises file to an output format such as html.
}
\examples{
\dontrun{render_exs(my_exs)}
}
