% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize-csv.R
\name{sanitize_csv}
\alias{sanitize_csv}
\title{sanitize csv}
\usage{
sanitize_csv(
  submission_file,
  path_to_submissions = "submissions/",
  replace_na_preds = NULL,
  alternative_name_for_pred = NULL,
  max_row = NULL,
  start_id = 1,
  name_id_var = "id",
  name_pred_column = "pred",
  verbose = TRUE
)
}
\arguments{
\item{submission_file}{name of csv file with predictions (chr)}

\item{path_to_submissions}{path to submission folder with submission files (chr)}

\item{replace_na_preds}{value to replace NA predictions (dbl)}

\item{alternative_name_for_pred}{alternative names for prediction column (str)}

\item{max_row}{how many rows should be prepared maximally (int)?}

\item{start_id}{number of the first id (int)}

\item{name_id_var}{name of the id variable (chr)?}

\item{name_pred_column}{name of the columns with the predictions (chr)?}

\item{verbose}{more output (lgl)?}
}
\value{
sanitized data frame of individual submission. See details.
}
\description{
Sanitizes a submission csv file for subsequent computation of prediction error
}
\details{
Takes a submission file as as input,
and irons out some glitches so that the prediction error can be computed
in a subsequent step (not part of this function).
Each single submissions consists of a number of submissions.
The submssions are expected to be a data frame with two columns:
\enumerate{
\item id: the unique id of each prediction (integer, positive)
\item pred: the predicted value for each instance
Note that the id must be shared knowledge,
i.e., the student must submit an id that is used by the teacher,
otherwise the predictions cannot be matched.
If the two column names are not "id" and "pred", the function
assumes the first and last column to be the relevant ones.
If only one column is provided as submission data frame,
the function assumes that the column contains the predictions. An id column is
added in this case.
Any missing values in the prediction will be filled with a given value,
provided the argument \code{replace_na_preds} is not NULL.
The submission does not necessarily be of a standard csv format, as
\code{data.table::fread} is able to guess formats. However, for the train and test data file, a standard csv file is expected (comma as deliminators, US centric locale).
}

The function returns a data frame with the the columns mentioned above (id, pred).
The returned dataframes contains some attributes: (1) "comments_to_student", where hints about gross errors are mentioned, (2) "fail" to indicate pass/fail of the test, here if gross errors (eg., no data submitted) are present, (3), "na_prop": the proportion of na in the prediction column.
}
\examples{
\dontrun{prep_csv(my_submission_file, path_train, path_test, max_row = 500, start_id = 800)}
}
