% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-teaching_data.R
\docType{data}
\name{teaching_data}
\alias{teaching_data}
\title{Datasets for teaching}
\format{
A dataframe with 12 rows
and 8 cols
}
\usage{
teaching_data
}
\description{
Datasets for teaching
}
\examples{
teaching_df

}
\keyword{datasets}
