% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_df.R
\name{teaching_df}
\alias{teaching_df}
\title{providing some data sets suitable for teaching}
\usage{
teaching_df(df = NULL, canonicalize = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{df}{name of dataframe to return, leave empty to get overview of available data sets. Enter "random" to get a randomly chosen data set.}
}
\value{
either dataframe of requested name or dataframe with overview of available data
}
\description{
Data sets suitable for (undergraduate) teaching should provide
variables at different scale leves (nominal, metric),
and should not be too complex or large.
Note that internet access is neccessary.
}
\examples{
\dontrun{teaching_df(df = "tips")}
}
