% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_readcsv.R
\name{try_readcsv}
\alias{try_readcsv}
\title{try_readcsv}
\usage{
try_readcsv(file, verbose = FALSE)
}
\arguments{
\item{file}{path to a csv file to be read (chr)}

\item{verbose}{more output to be printed (lgl)?}
}
\value{
tibble
}
\description{
tries to rad a submission csv file
}
\examples{
\dontrun{try_readcsv(my_submission_file)}
}
