% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_by_two_association.R
\name{two_by_two_association}
\alias{two_by_two_association}
\title{Create 2x2 data set from teaching data}
\usage{
two_by_two_association(data = "random")
}
\arguments{
\item{data}{the dataset the used}
}
\value{
Dataframe with two variables are factors,
where the labels are "high" and "low"
}
\description{
This function creates a two-by-two assocation type dataset
from the list of teaching data sets from this package
}
\details{
If \code{data} is set to "random" a randomly chosen data set
fromt the list of available data sets is chosen.
Alternatively, one name can be specified.
See \code{teaching_data} for details.
Then, the outcome variable and the first predictors are
split at the median.
}
\examples{
two_by_two_data <- two_by_two_association("random")
}
