% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yamlrmdfile.R
\name{write_yamlrmdfile}
\alias{write_yamlrmdfile}
\title{write_yamlrmdfile}
\usage{
write_yamlrmdfile(
  ex_parsed,
  path_output,
  ex_sol_str = c("Exercise", "Solution", "Categories"),
  print_categories = TRUE,
  separate_ex_sol = rep("</br>", 10),
  verbose = TRUE,
  more_yaml = NULL,
  header_level = 1
)
}
\arguments{
\item{ex_parsed}{list output given by parse_examfile()}

\item{path_output}{where should the output file saved to?}

\item{ex_sol_str}{wording for the parts of the exercise (useful for translation purposes)}

\item{print_categories}{should the yaml field 'categories' be printed?}

\item{separate_ex_sol}{How should the question and solution be visually separated?}

\item{verbose}{talkative, yes or no? (TRUE or FALSE)}

\item{more_yaml}{more yaml to add which was not given in the r-exam file}

\item{header_level}{how many '#' should be padded in order to indicate the header leve?}
}
\value{
writes an rmd file
}
\description{
Takes parts of r-exam files and writes into a yaml rmd file
}
\details{
Take the parts of r-exam files which must be handed over
by the function parse_examfile() and then writes rmd file with yaml header.
}
\examples{
\dontrun{write_yamlrmdfile(ex_parsed, my_path)}
}
