% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_grades.R
\name{assign_grade}
\alias{assign_grade}
\title{Assign Grades}
\usage{
assign_grade(d, grading_scheme = NULL, var = "correct_prop")
}
\arguments{
\item{d}{data frame}

\item{grading_scheme}{which grading scheme to use}

\item{var}{variable for which a grade should be assigned}
}
\value{
data frame with grades (column) added
}
\description{
Given a variable with points, percentage or other, a grade is
assigned given a grading scheme which acts as a function ascribing the
mapping from proints to grade
}
\examples{
assign_grade(mtcars, var = "am")
}
