% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam2yamlrmd.R
\name{exam2yamlrmd}
\alias{exam2yamlrmd}
\title{exam2yamlrmd}
\usage{
exam2yamlrmd(
  examfile,
  path_output,
  ex_sol_str = c("Exercise", "Solution", "Categories"),
  print_categories = FALSE,
  separate_ex_sol = rep("</br>", 10),
  verbose = TRUE,
  header_level = 1
)
}
\arguments{
\item{examfile}{input file in r-exam format}

\item{path_output}{output puath}

\item{ex_sol_str}{wording for the parts of the exercise (useful for translation purposes)}

\item{print_categories}{should the yaml field 'categories' be printed?}

\item{separate_ex_sol}{How should the question and solution be visually separated?}

\item{verbose}{talkative, yes or no? (TRUE or FALSE)}

\item{header_level}{how many '#' should be padded in order to indicate the header leve?}

\item{more_yaml}{more yaml to add which was not given in the r-exam file}
}
\value{
writes an rmd file
}
\description{
Converts r-exam File to yaml-headed RMD file
}
\details{
This function calls parse_examfile() and then write_yamlrmdfile().
}
\examples{
\dontrun{exam2yamlrmd(ex_parsed, my_path)}
}
