% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_em.R
\name{grade_em}
\alias{grade_em}
\title{Grade em}
\usage{
grade_em(x, breaks12, reverse = FALSE)
}
\arguments{
\item{x}{student's percentage points (one row per student)}

\item{reverse}{reverse grading scheme (lgl)?}

\item{breaks}{thresholds for a given grade (vector of length 12)}
}
\value{
grades per student
}
\description{
This function assigns grades
}
\examples{
\dontrun{grade_em(thresholds, values)}
}
