% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-funs.R
\name{grade_em}
\alias{grade_em}
\title{Grade em}
\usage{
grade_em(thresholds, values)
}
\arguments{
\item{thresholds}{thresholds for a given grade}

\item{values}{student's percentage points (one row per student)}
}
\value{
grades per student
}
\description{
This function assigns grades
}
\examples{
\dontrun{grade_em(thresholds, values)}
}
