% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_em.R
\name{grade_em}
\alias{grade_em}
\title{Grade em}
\usage{
grade_em(x, thresholds, reverse = FALSE, grades_scheme = NULL, verbose = FALSE)
}
\arguments{
\item{x}{student's prediction error value, such as MAE or RMSE (numeric)}

\item{thresholds}{thresholds for a given grade (numeric vector of length 12)}

\item{reverse}{reverse grading scheme (lgl)?}

\item{grades_scheme}{Grades scheme, eg., from 5.0 to 1.0}
}
\value{
grades per student
}
\description{
This function assigns grades
}
\details{
If no grade scheme is given, the function uses it own grade scheme,
which is printed when the argument \code{verbose} is \code{TRUE}.
The number of thresholds ("fence posts") must be equal the number of grades ("fence ways") plus 1.
The default grade \code{grades_scheme} used is  \code{c(5, 4, 3.7, 3.3, 3.0, 2.7, 2.3, 2, 1.7, 1.3, 1)},
which is of length 11
´
}
\examples{
\dontrun{grade_em(thresholds, values)}
}
