% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_em_all.R
\name{grade_em_all}
\alias{grade_em_all}
\title{Grade em all}
\usage{
grade_em_all(
  data_path_moodle_csv,
  grading_scheme = NULL,
  verbose = TRUE,
  n_bonus_points = 0,
  comments = NULL
)
}
\arguments{
\item{data_path_moodle_csv}{A csv file, possible coming from moodle with exam items. See details.}

\item{grading_scheme}{Defines thresholds for grades}

\item{verbose}{if true, additional processing infos are printed}

\item{n_bonus_points}{Be kind to your students}

\item{comments}{comments passed on to each student}
}
\value{
csv file with grades
}
\description{
Assigns grades to exams
}
\details{
This function assigns a grade to each student (row)
drawing upon the exam items in the input csv moodle file.
Some preprocssing is being done, such as cleaning the col names.
}
\examples{
\dontrun{grade_em_all(data.csv)}

}
