% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_gradinglists.R, R/join_gradinglists3.R
\name{join_gradinglists}
\alias{join_gradinglists}
\title{Join two grading lists}
\usage{
join_gradinglists(
  moodle_grading_d,
  course_file,
  comments = NULL,
  grades_var = "grade"
)

join_gradinglists(
  moodle_grading_d,
  course_file,
  comments = NULL,
  grades_var = "grade"
)
}
\arguments{
\item{moodle_grading_d}{data frame with results of moodle grading}

\item{course_file}{path to XLSX template file from university}

\item{comments}{comments passed to each student (separate comments are not supported)}

\item{grades_var}{variable in Moodle table holding the grades of the students}
}
\value{
data frame for grading based on the institution name list

data frame for grading based on the instutation name list
}
\description{
Join moodle grading list with Instituation's grading list

Join moodle grading list with Instituation's grading list
}
\details{
Based on the Moodle csv file, a grading list has been created.
This function now merges the moodle grading list with a XLSX template file
from the institution. Merging (joining) is performed by based on surname
and first names of the student. Joining errors are reported. Note that
names are no unique nor save way to join lists.
ATTENTION:

Based on the Moodle csv file, a grading list has been created.
This function now merges the moodle grading list with a XLSX template file
from the institution. Merging (joining) is performed by based on surname
and first names of the student. Joining errors are reported. Note that
names are no unique nor save way to join lists.
}
\examples{
 \dontrun{join_gradinglists(d1, d2)}
 \dontrun{join_gradinglists(d1, d2)}
}
