% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_notenspiegel.R
\name{plot_notenspiegel}
\alias{plot_notenspiegel}
\title{Plot Notenspiegel}
\usage{
plot_notenspiegel(d, plot_title = "Notenverteilung", var = "grade")
}
\arguments{
\item{d}{data frame containing grades}

\item{plot_title}{Title to be assigned to the plot}

\item{var}{variable holding the grades to be plotted}
}
\value{
ggplot2 object
}
\description{
Plot Grade Distribution (Notenspiegel).
}
\details{
Plot distribution of grades, a simple wrapper around ggplot2
Annotations are currently in German language
}
\examples{
\dontrun{plot_notenspiegel(grades_data)}
}
