% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_notenspiegel.R
\name{plot_notenspiegel}
\alias{plot_notenspiegel}
\title{Plot Notenspiegel}
\usage{
plot_notenspiegel(
  d,
  plot_title = "Notenverteilung",
  var = "grade",
  show_prop_avg = FALSE
)
}
\arguments{
\item{d}{data frame containing grades}

\item{plot_title}{Title to be assigned to the plot}

\item{var}{variable holding the grades to be plotted}

\item{show_prop_avg}{Should the mean percentage value be shown (defaults to FALSE)}
}
\value{
ggplot2 object
}
\description{
Plot Grade Distribution (Notenspiegel).
}
\details{
Plot distribution of grades, a simple wrapper around ggplot2
Annotations are currently in German language

IF \code{show_prop_avg} is TRUE, it is expected that a column by the name of \code{prop} is available which holds the percentage of points achieved by student
}
\examples{
\dontrun{plot_notenspiegel(grades_data)}
}
